Front: Split in modules
