+++ rts-front/src/screens/mod.rs
@@ -1,0 +1,2 @@
+pub mod login;
+pub mod ai_upload;
